requirements.txt
python-dotenv>=0.17.1<0.18.0
Pillow>=8.3.2<8.4.0
Flask>=2.0.2<2.1.0
waitress>=2.0.0<2.1.0